+++ src/Cardian.ts
@@ -1,19 +1,18 @@
+import type * as types from "/types";
 import * as common from "/common";
 import * as blackjack from "/blackjack";
-// import * as holdem from "/holdem";
+import * as holdem from "/holdem";
 
 import * as decklists from "/decklists";
 
-interface IComponents {
+export interface IComponents {
 	Card: Card;
 	Box: Box;
 	List: List;
 }
-type Card = typeof common.Card | typeof blackjack.Card; // | typeof holdem.Card;
-type Box = typeof common.Box | typeof blackjack.Box; // | typeof holdem.Card;
-type List = typeof common.List | typeof blackjack.List; // | typeof holdem.Card;
-
-type Mode = "none" | "blackjack"; // | "holdem";
+export type Card = typeof common.Card | typeof blackjack.Card | typeof holdem.Card;
+export type Box = typeof common.Box | typeof blackjack.Box | typeof holdem.Box;
+export type List = typeof common.List | typeof blackjack.List | typeof holdem.List;
 
 class Cardian {
 	static singleton = false;
@@ -29,11 +28,11 @@ class Cardian {
 		return this.#components.List;
 	}
 
-	#standard52: common.PseudoCard[];
+	#standard52: types.PseudoCard[];
 	get standard52() {
 		return this.#standard52;
 	}
-	#standard52_sextet: common.PseudoCard[];
+	#standard52_sextet: types.PseudoCard[];
 	get standard52_sextet() {
 		return this.#standard52_sextet;
 	}
@@ -50,7 +49,7 @@ class Cardian {
 		this.#standard52_sextet = decklists.standard52_sextet;
 	}
 
-	setMode(mode: Mode = "none"): void {
+	setMode(mode: types.Mode = "none"): void {
 		switch (mode) {
 			case "none":
 				this.#components = common;
@@ -60,9 +59,9 @@ class Cardian {
 				this.#components = blackjack;
 				break;
 
-			// case "holdem":
-			//     this.#components = holdem;
-			// 	break;
+			case "holdem":
+				this.#components = holdem;
+				break;
 		}
 	}
 }