+++ src/holdem/Box.ts
@@ -1,0 +1,9 @@
+import { Box } from "/common";
+
+class HEBox extends Box {
+	constructor(name: string) {
+		super(name);
+	}
+}
+
+export { HEBox as Box };