+++ src/holdem/List.ts
@@ -1,0 +1,10 @@
+import type { PseudoCard } from "/types";
+import { List } from "/common";
+
+class HEList extends List {
+	constructor(pseudoCards: Array<PseudoCard>) {
+		super(pseudoCards);
+	}
+}
+
+export { HEList as List };